refactor(errors): move `ParseException` to `kage.errors` package
