+++ src/kotlin/kage/format/AgeStanza.kt
@@ -8,6 +8,8 @@ package kage.format
 import java.io.BufferedReader
 import java.io.BufferedWriter
 import java.util.Base64
+import kage.errors.InvalidArbitraryStringException
+import kage.errors.InvalidRecipientException
 import kage.format.AgeKey.Companion.BYTES_PER_LINE
 import kage.format.AgeKey.Companion.COLUMNS_PER_LINE
 import kage.format.AgeKey.Companion.FOOTER_PREFIX
